Refactor TabsPage and charts: rename components, update imports, and add UserActivityHeatmap and UserDemographicsChart
