+++ web-ui/src/components/tabs/TabGrowthChart.jsx
